Include error body in inner exception
When an inner exception is not provided, include the response body in
the FlurlHttpException's inner exception. This allows most
application's default exception handling to have access to the error
details.

For example, unit test frameworks and loggers will capture both the
root and any inner exceptions but won't capture the HttpCall (and error
body). This enables tests to capture/display the error body without
having to wrap every test in a try/catch.