Test Assertion WithContentType no longer throws NullReferenceException if no content was sent
If a some requests were made to a URL containing the same part where some requests included
content and some did not, attempting to assert that the client ShouldHaveCalled that part
with a post and a contnet type would throw a NullReferenceException when it encountered
the requests that did not have any content sent with them. Ignore requests with no content
when asserting on ContentType.

Fixes #285