Add tooling for building only nestandard and netcoreapp targets
This is in prep for #286 and allows users on macOS and Linux to build
and test only the netstandard and netcoreapp targets. All published
projects support at least netstandard1.1 which is consumable from net45,
so if we decide to just have everyone consume a netstandard package this
can be greatly simplified.